+++ samples/Sentry.Samples.Android/C/buggy.c
@@ -1,0 +1,6 @@
+__attribute__((noinline))
+void crash_in_c()
+{
+    char *ptr = 0;
+    *ptr += 1;
+}