+++ src/Sentry.DiagnosticSource/Internals/DiagnosticSource/SentryDiagnosticListenerIntegration.cs
@@ -1,11 +1,11 @@
 using System;
 using System.Diagnostics;
 using Sentry.Extensibility;
-using Sentry.Internal;
+using Sentry.Integrations;
 
 namespace Sentry.Internals.DiagnosticSource
 {
-    internal class SentryDiagnosticListenerIntegration : IInternalSdkIntegration
+    internal class SentryDiagnosticListenerIntegration : ISdkIntegration
     {
         private SentryDiagnosticSubscriber? _subscriber;
         private IDisposable? _diagnosticListener;
@@ -21,12 +21,6 @@ namespace Sentry.Internals.DiagnosticSource
 
             _subscriber = new SentryDiagnosticSubscriber(hub, options);
             _diagnosticListener = DiagnosticListener.AllListeners.Subscribe(_subscriber);
-        }
-
-        public void Unregister(IHub hub)
-        {
-            _diagnosticListener?.Dispose();
-            _subscriber?.Dispose();
         }
     }
 }