+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Features
 
 - Initial support for .NET MAUI ([#1663](https://github.com/getsentry/sentry-dotnet/pull/1663))
+### Fixes
+
+- Remove IInternalSdkIntegration ([#1656](https://github.com/getsentry/sentry-dotnet/pull/1656))
+- On async Main, dont unregister unhandled exception before capturing crash  ([#321](https://github.com/getsentry/sentry-dotnet/issues/321))
 
 ## 3.17.1
 
@@ -21,6 +25,7 @@
 
 - Collect and send Client Reports to Sentry, which contain counts of discarded events. ([#1556](https://github.com/getsentry/sentry-dotnet/pull/1556))
 - Expose `ITransport` and `SentryOptions.Transport` public, to support using custom transports ([#1602](https://github.com/getsentry/sentry-dotnet/pull/1602))
+- Android native crash support ([#1288](https://github.com/getsentry/sentry-dotnet/pull/1288))
 
 ### Fixes